refactor : Seat 관련 클래스 이름 변경 및 일부 위치 변경 (2)
