+++ src/main/java/org/example/siljeun/domain/reservation/service/ReservationService.java
@@ -8,7 +8,7 @@ import org.example.siljeun.domain.reservation.entity.Reservation;
 import org.example.siljeun.domain.reservation.exception.CustomException;
 import org.example.siljeun.domain.reservation.exception.ErrorCode;
 import org.example.siljeun.domain.reservation.repository.ReservationRepository;
-import org.example.siljeun.domain.seat.repository.SeatScheduleInfoRepository;
+import org.example.siljeun.domain.seatscheduleinfo.repository.SeatScheduleInfoRepository;
 import org.example.siljeun.domain.seat.entity.SeatScheduleInfo;
 import org.example.siljeun.domain.seat.enums.SeatStatus;
 import org.example.siljeun.domain.user.entity.User;
