+++ src/main/java/org/example/siljeun/domain/seat/repository/SeatRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface VenueSeatRepository extends JpaRepository<Seat, Long> {
+public interface SeatRepository extends JpaRepository<Seat, Long> {
 
     List<Seat> findByVenue(Venue venue);
 }
